Add Refinement#refined_class
