+++ src/main/java/org/truffleruby/core/module/ModuleNodes.java
@@ -32,7 +32,6 @@ import com.oracle.truffle.api.profiles.InlinedLoopConditionProfile;
 import com.oracle.truffle.api.strings.TruffleString;
 import com.oracle.truffle.api.strings.TruffleString.ByteIndexOfStringNode;
 import org.truffleruby.RubyContext;
-import org.truffleruby.RubyLanguage;
 import org.truffleruby.annotations.CoreMethod;
 import org.truffleruby.builtins.CoreMethodArrayArgumentsNode;
 import org.truffleruby.builtins.CoreMethodNode;
@@ -2466,73 +2465,6 @@ public abstract class ModuleNodes {
         }
 
     }
-
-    @Primitive(name = "refinement_import_methods")
-    public abstract static class ImportMethodsNode extends PrimitiveArrayArgumentsNode {
-
-        @TruffleBoundary
-        @Specialization
-        protected RubyModule importMethods(RubyModule refinement, RubyModule moduleToImportFrom) {
-            var firstNonRubyMethod = getFirstNonRubyMethodOrNull(moduleToImportFrom, getLanguage());
-            if (firstNonRubyMethod != null) {
-                throw new RaiseException(getContext(),
-                        coreExceptions().argumentError(createErrorMessage(firstNonRubyMethod, moduleToImportFrom),
-                                this));
-            }
-
-            importMethodsFromModuleToRefinement(moduleToImportFrom, refinement);
-
-            return refinement;
-        }
-
-        private String createErrorMessage(InternalMethod method, RubyModule module) {
-            return StringUtils.format("Can't import method which is not defined with Ruby code: %s#%s",
-                    module.getName(), method.getName());
-        }
-
-        private void importMethodsFromModuleToRefinement(RubyModule module, RubyModule refinement) {
-            var declarationContext = createDeclarationContextWithRefinement(refinement);
-            for (InternalMethod methodToCopy : module.fields.getMethods()) {
-                var clonedMethod = cloneMethod(methodToCopy, declarationContext, refinement);
-                refinement.fields.addMethod(getContext(), this, clonedMethod);
-            }
-        }
-
-        private InternalMethod getFirstNonRubyMethodOrNull(RubyModule module, RubyLanguage language) {
-            for (InternalMethod method : module.fields.getMethods()) {
-                if (!method.isDefinedInRuby(language)) {
-                    return method;
-                }
-            }
-
-            return null;
-        }
-
-        // Creates a declaration context which contains the refined methods from the given refinement
-        private DeclarationContext createDeclarationContextWithRefinement(RubyModule refinement) {
-            final Map<RubyModule, RubyModule[]> refinements = new HashMap<>();
-            refinements.put(refinement.fields.getRefinedModule(), new RubyModule[]{ refinement });
-            return new DeclarationContext(
-                    Visibility.PUBLIC,
-                    new FixedDefaultDefinee(refinement),
-                    refinements);
-        }
-
-        private InternalMethod cloneMethod(InternalMethod method, DeclarationContext declarationContext,
-                RubyModule refinement) {
-            var clonedCallTarget = cloneCallTarget(method);
-            return method.withCallTargetAndDeclarationContextAndDeclarationModule(clonedCallTarget, declarationContext,
-                    refinement);
-        }
-
-        private RootCallTarget cloneCallTarget(InternalMethod method) {
-            var rubyRootNode = (RubyRootNode) method.getCallTarget().getRootNode();
-            var clonedRootNode = rubyRootNode.cloneUninitialized();
-
-            return clonedRootNode.getCallTarget();
-        }
-    }
-
 
     @GenerateUncached
     @CoreMethod(names = "using", required = 1, visibility = Visibility.PRIVATE, alwaysInlined = true)
