+++ test/mri/tests/ruby/test_refinement.rb
@@ -1835,6 +1835,20 @@ class TestRefinement < Test::Unit::TestCase
     assert_equal [ref::RefB, ref::RefA], ref::Combined::USED_MODS
   end
 
+  def test_refined_class
+    refinements = Module.new {
+      refine Integer do
+        int_refinement = self
+      end
+
+      refine String do
+        str_refinement = self
+      end
+    }.refinements
+    assert_equal(Integer, refinements[0].refined_class)
+    assert_equal(String, refinements[1].refined_class)
+  end
+
   def test_warn_setconst_in_refinmenet
     bug10103 = '[ruby-core:64143] [Bug #10103]'
     warnings = [
